Update README.md
<h2>Description</h2>

To create a password generator in which the user has to provide a range of characters between 8 and 128. The user will have to choose at least one of the given criterias: lower case, upper case, numbers & characters. If user decides to not follow any of the given prompts the wite will instruct the user to meet the criterias.

<h4> Link: </h4>

* [Github Link](https://github.com/channy3687/channy3687.password-generator)
* [Horiseon website](https://channy3687.github.io/channy3687.Horiseo/)</h2>

The project is a code refactor for the Horiseon website. With refractoing the page it will make the code more efficent, accesible, and easily understandable without changing the page's code or style. 

<h3> Built With </h3>

* HTML
* CSS

<h4> Link: </h4>

* [Github Link](https://github.com/channy3687/channy3687.Horiseo.git)
* [Horiseon website](https://channy3687.github.io/channy3687.Horiseo/)